+++ src/main/java/com/bob/jr/interfaces/VoidCommand.java
@@ -1,0 +1,7 @@
+package com.bob.jr.interfaces;
+
+import reactor.core.publisher.Mono;
+
+public interface VoidCommand {
+    Mono<Void> execute();
+}
